// src/logic/rest/api.ts
export interface Body2 {
    /**
     * The new value for the field `firstname`
     * @type {string}
     * @memberof Body2
     */
    firstname?: string;
    /**
     * The new value for the field `lastname`
     * @type {string}
     * @memberof Body2
     */
    lastname?: string;
    /**
     * The new value for the field `email`
     * @type {string}
     * @memberof Body2
     */
    email?: string;
    /**
     * The new password
     * @type {string}
     * @memberof Body2
     */
    password?: string;
}
// ...
export interface User {
    /**
     * The id of the user
     * @type {number}
     * @memberof User
     */
    id: number;
    /**
     * The username
     * @type {string}
     * @memberof User
     */
    username: string;
    /**
     * First name of the user
     * @type {string}
     * @memberof User
     */
    firstname: string;
    /**
     * Last name of the user (only visible to admins)
     * @type {string}
     * @memberof User
     */
    lastname?: string;
    /**
     * E-mail of the user (only visible to admins)
     * @type {string}
     * @memberof User
     */
    email?: string;
    /**
     * `true` if the email is verified, else `false`
     * @type {boolean}
     * @memberof User
     */
    emailVerified: boolean;
}